requirements.txt
streamlit==1.32.0
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.12.0
pandas==2.2.0
python-dotenv==1.0.1
redis==5.0.1
